Reorganized Dijkstra
